+++ fpsim/__init__.py
@@ -1,4 +1,5 @@
 from .version import __version__, __versiondate__
+from .settings import *
 from .utils import *
 from .defaults import *
 from .parameters import *