+++ tests/test_scenarios.py
@@ -9,7 +9,7 @@ import pytest
 
 # Global settings
 int_year = 2002 # Year to start the interventions
-serial   = 1 # Whether to run in serial (for debugging)
+serial   = 0 # Whether to run in serial (for debugging)
 do_plot  = 1 # Whether to do plotting in interactive mode
 sc.options(backend='agg') # Turn off interactive plots
 
@@ -189,9 +189,8 @@ def test_scenarios(do_plot=do_plot):
 
     # Plot and print results
     if do_plot:
+        scens.plot()
         scens.plot_sims()
-        scens.plot_scens()
-        scens.plot_cpr()
         scens.plot_method_mix()
 
     return scens
@@ -212,7 +211,7 @@ def test_make_scens():
     s.eff   = fp.make_scen(year=year, eff={'Injectables':0.99}) # Basic efficacy scenario
     s.prob1 = fp.make_scen(year=year, source='None', dest='Injectables', factor=2) # Double rate of injectables initiation
     s.prob2 = fp.make_scen(year=year, method='Injectables', init_factor=2) # Double rate of injectables initiation -- alternate approach
-    s.par   = fp.make_scen(par='exposure_factor', years=2005, vals=0.5) # Parameter scenario: halve exposure
+    s.par   = fp.make_scen(par='exposure_factor', par_years=2005, par_vals=0.5) # Parameter scenario: halve exposure
 
     # More complex example: change condoms to injectables transition probability for 18-25 postpartum women
     s.complex = fp.make_scen(year=year, source='Condoms', dest='Injectables', value=0.5, ages='18-20', matrix='pp1to6')
@@ -222,20 +221,18 @@ def test_make_scens():
     s.custom = fp.make_scen(interventions=update_sim)
 
     # Combining multiple scenarios: increase injectables initiation and reduce exposure factor
-    s.multi = fp.make_scen(
+    s.multi = fp.make_scen([
         dict(year=year, method=method, init_factor=2),
-        dict(par='exposure_factor', years=2010, vals=0.5)
-    )
+        dict(par='exposure_factor', par_years=2010, par_vals=0.5)
+    ])
 
     # Scenario addition
     s.sum = s.eff + s.prob1
 
     # More probability matrix options
     s.inj1 = fp.make_scen(year=year, method=method, init_factor=5, matrix='annual', ages=None)
-    s.inj2 = fp.make_scen(year=year, method=method, discont_factor=0, matrix='annual', ages=':')
-    s.inj3 = fp.make_scen(year=year, method=method, init_value=0.2, matrix='pp1to6', ages=None)
-    s.inj4 = fp.make_scen(year=year, method=method, discont_value=0, matrix='pp1to6', ages=':')
-    s.inj5 = fp.make_scen(year=year, source='None', dest='Injectables', factor=0.2, ages=['<18', '>25'])
+    s.inj2 = fp.make_scen(year=year, method=method, discont_value=0, matrix='pp1to6', ages=':')
+    s.inj3 = fp.make_scen(year=year, source='None', dest='Injectables', factor=0.2, ages=['<18', '>25'])
 
     # Test invalid options
     with pytest.raises(sc.KeyNotFoundError):